ton of refactor. hard to keep track.
