--- .env
@@ -1,4 +1,0 @@
-DEV=1
-
-# go run test/github_server/main.go
-# URL_LATEST_RELEASE=http://localhost:5678/releases/latest
