--- cmd/playground/playground.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	"log"
-)
-
-func main() {
-	log.SetFlags(log.Default().Flags() | log.Lshortfile)
-}
