+++ go.mod
@@ -1,5 +1,3 @@
 module github.com/mysterion/avrp
 
-go 1.22.3
-
-require github.com/rs/cors v1.11.0 // indirect
+go 1.23
