+++ internal/server/init.go
@@ -1,4 +1,4 @@
-package api
+package server
 
 var servDir string
