+++ internal/server/server.go
@@ -1,0 +1,166 @@
+package server
+
+import (
+	"context"
+	"crypto/tls"
+	"embed"
+	"fmt"
+	"log"
+	"net"
+	"net/http"
+	"os"
+	"os/signal"
+	"time"
+
+	"github.com/mysterion/avrp/internal/utils"
+)
+
+//go:embed certs/*
+var certs embed.FS
+
+func getIps() []string {
+	if utils.IsGoRun() {
+		return []string{"127.0.0.1"}
+	}
+	ips := make([]string, 0)
+	ifaces, err := net.Interfaces()
+	if err != nil {
+		panic(err)
+	}
+	for _, i := range ifaces {
+		addrs, err := i.Addrs()
+		if err != nil {
+			panic(err)
+		}
+		for _, addr := range addrs {
+			var ip net.IP
+			switch v := addr.(type) {
+			case *net.IPNet:
+				ip = v.IP
+			case *net.IPAddr:
+				ip = v.IP
+			}
+			if ip.To4() != nil && (ip.IsPrivate() || ip.IsLoopback()) {
+				ips = append(ips, ip.String())
+			}
+		}
+	}
+	return ips
+}
+
+func getTlsConfig() (*tls.Config, error) {
+	config := &tls.Config{
+		MinVersion:               tls.VersionTLS12,
+		PreferServerCipherSuites: true,
+		Certificates:             make([]tls.Certificate, 1),
+	}
+
+	certBlock, err := certs.ReadFile("certs/cert.pem")
+	if err != nil {
+		return nil, err
+	}
+
+	keyBlock, err := certs.ReadFile("certs/key.pem")
+	if err != nil {
+		return nil, err
+	}
+
+	config.Certificates[0], err = tls.X509KeyPair(certBlock, keyBlock)
+	if err != nil {
+		return nil, err
+	}
+
+	return config, nil
+}
+
+// port = 0, for a random port
+//
+// mux can be nil
+//
+// tlsConfig can be nil
+//
+// returns [<-ready], [<-done], [*http.Server]
+func new(port int, mux http.Handler, tlsConfig *tls.Config) (<-chan bool, <-chan bool, *http.Server) {
+
+	ready := make(chan bool, 1)
+	done := make(chan bool, 1)
+
+	server := http.Server{
+		Handler:   mux,
+		TLSConfig: tlsConfig,
+	}
+
+	go func() {
+		defer close(ready)
+		defer close(done)
+
+		url := "0.0.0.0"
+		if utils.IsGoRun() {
+			url = "127.0.0.1"
+		}
+
+		listener, err := net.Listen("tcp4", fmt.Sprintf("%s:%d", url, port))
+		utils.Panic(err)
+
+		server.Addr = listener.Addr().String()
+
+		ready <- true
+
+		err = server.ServeTLS(listener, "", "")
+
+		if err != nil && err != http.ErrServerClosed {
+			fmt.Println(err.Error())
+			panic(err)
+		}
+	}()
+
+	return ready, done, &server
+}
+
+func Start(port int) {
+	tlsConfig, err := getTlsConfig()
+	if err != nil {
+		panic(err)
+	}
+
+	mux := http.NewServeMux()
+
+	const filePath = "/file/"
+	var fileHandler = http.StripPrefix(filePath, http.FileServer(http.Dir(servDir)))
+	mux.Handle(filePath, fileHandler)
+
+	const distPath = "/"
+	var distHandler = http.FileServer(http.Dir(utils.DistDir))
+	mux.Handle(distPath, distHandler)
+
+	mux.HandleFunc(listPath, listHandler)
+	mux.HandleFunc(thumbPath, thumbHandler)
+
+	sigint := make(chan os.Signal, 1)
+	signal.Notify(sigint, os.Interrupt)
+
+	ready, done, s := new(port, mux, tlsConfig)
+	<-ready
+
+	fmt.Println("Server listening on: ")
+	for _, ip := range getIps() {
+		fmt.Printf("https://%v:%v\n", ip, port)
+	}
+	<-sigint
+
+	ctx, cancel := context.WithDeadline(context.TODO(), time.Now().Add(time.Second*3))
+	defer cancel()
+
+	s.Shutdown(ctx)
+
+	log.Println("Shutting down..")
+
+	select {
+	case <-done:
+		log.Println("bye")
+	case <-ctx.Done():
+		log.Println("shutdown request timedout..")
+		log.Println("ok..")
+	}
+
+}
