--- web/api/certs/cert.pem
@@ -1,33 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIFnTCCA4WgAwIBAgIUNmG+My915s9o1EoSwLG0uyxgyFcwDQYJKoZIhvcNAQEL
-BQAwXjELMAkGA1UEBhMCQVUxEzARBgNVBAgMClNvbWUtU3RhdGUxEjAQBgNVBAoM
-CW15c3RlcmlvbjELMAkGA1UECwwCVlIxGTAXBgNVBAMMEGFmcmFtZS12ci1wbGF5
-ZXIwHhcNMjQwNTEwMDg0NzE0WhcNMzQwNTA4MDg0NzE0WjBeMQswCQYDVQQGEwJB
-VTETMBEGA1UECAwKU29tZS1TdGF0ZTESMBAGA1UECgwJbXlzdGVyaW9uMQswCQYD
-VQQLDAJWUjEZMBcGA1UEAwwQYWZyYW1lLXZyLXBsYXllcjCCAiIwDQYJKoZIhvcN
-AQEBBQADggIPADCCAgoCggIBAKnR4NheDVcLqv3jBSBmtmSud9KrRUuJ4fppJp/M
-7n2FVZvxKwKaWsIW5xeNwAW5vbQ2lhYixSPaxhKniOsociek6drDLUnN1DyWHngz
-lZpxkzkUl4WF6Nh+V2oYNSMNr0YI74Zb/jFPMZDTqNs7761iRH6i/aWlGmCJQm3U
-Ra6U8J4LuMT0tNZLC4tT8GAxEIVg3eUXtahoaslMhsAzMmEPIGzPP+vecNOVrKwo
-noeIzT0tYWhgtWdYe9ayjg/R2+ioBIBUaxt9FsDTDCvfOHAVQ0lluFU/5uCcESOI
-w8QM9kpvfj6vlEUweys1ycRRxRcus/SOMShhOtt7sCoK+IU3RLSOlzYZgjxuMpZi
-gm9N8ebYrVXPRSfF4evwdU46k6Xx1nOMCw8KwT2aqEYif7ImFIdik0/5+TB47Ezr
-66Bwe8Cu0NFdyiQz8MvASPGGkAlPKUOO+8wQJCkFc1nWnkqATQ5IyvmSgj/dVc9W
-oBmKy6aArDEEOrj8TucbJ/J6IuUGPGPMQJCzasWewRav7JZ97LWn96PdwYejSM1t
-PLxzukbPZ/o4IFVG/sbwC6UCDgofcEz5yjgqwF1BeKiXbi0CmOAGRYdfpc44lYIS
-SHNmfz0BW/i/Gu4wJerrucBZS/I0uUj9Qawwrf6BZkTJz5fvXs7YyC8iT4DGKmiu
-vxQjAgMBAAGjUzBRMB0GA1UdDgQWBBRC8T4jL8WzeBYu9+HYkFur79xFmTAfBgNV
-HSMEGDAWgBRC8T4jL8WzeBYu9+HYkFur79xFmTAPBgNVHRMBAf8EBTADAQH/MA0G
-CSqGSIb3DQEBCwUAA4ICAQByX2j0bUa8umQ43ZJd6Yr78yQ31ERcnL6dtWiYSspn
-7AkySTwq94oTKWPqKMbmjPwANHxRX5MopD/5y2dMVU+08z0hExgNk3AyFL43E4xT
-CRXGx7rcqhwzQRYwDsOWfcUFTi2yCA6x8C08Xh0gZ1kMBB4L8Ms2rzTEjTAAK2C/
-KK70htUj1jF4EZjkq/WX3qcaLxcn8lnuveRYu0RryVabxr+ENiBJvs+XinklUYIh
-8C32u70x2/uq7e3/D6ir811JDA2F0dFoGawjTnKRGFn6vpB1GkJvlwecgeemT829
-9/tzW9aoyEDqVxJbvfgyOY+UBraHRuWOy0gSWrgIKhPbQeZjJTW+sFOpyjxF1oN/
-QVnGAPkySCwInrwI6jol1jbWQzUmN8Zk12Gp+GWxMp0xP9v/OqfAPs+VwOn5OPQT
-SXB/dngDvrf58c7FljjUZU10/g+T6/Vh1M1NwDc8d55+Ogzc3GvD6msd0TSxmtGQ
-Bas75YYpEH1m8dCxn/hgQ1lcbYwdlVz0zlSI2rV3gHirPxDGlcWOy21DLpw79KCg
-hFl3gY1AKEGd7WhHAPISm4nLyW/V795jqVFCo2B08QMWYVBfocl2CgqUKEWU+REl
-pRYIqtCPNNjBpVpHDqY21nHx7PIa+lKLlhoQPzzGqUbHtefnml3a20LcTkjFeGS+
-VQ==
------END CERTIFICATE-----
