--- web/api/certs/key.pem
@@ -1,52 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIIJQgIBADANBgkqhkiG9w0BAQEFAASCCSwwggkoAgEAAoICAQCp0eDYXg1XC6r9
-4wUgZrZkrnfSq0VLieH6aSafzO59hVWb8SsCmlrCFucXjcAFub20NpYWIsUj2sYS
-p4jrKHInpOnawy1JzdQ8lh54M5WacZM5FJeFhejYfldqGDUjDa9GCO+GW/4xTzGQ
-06jbO++tYkR+ov2lpRpgiUJt1EWulPCeC7jE9LTWSwuLU/BgMRCFYN3lF7WoaGrJ
-TIbAMzJhDyBszz/r3nDTlaysKJ6HiM09LWFoYLVnWHvWso4P0dvoqASAVGsbfRbA
-0wwr3zhwFUNJZbhVP+bgnBEjiMPEDPZKb34+r5RFMHsrNcnEUcUXLrP0jjEoYTrb
-e7AqCviFN0S0jpc2GYI8bjKWYoJvTfHm2K1Vz0UnxeHr8HVOOpOl8dZzjAsPCsE9
-mqhGIn+yJhSHYpNP+fkweOxM6+ugcHvArtDRXcokM/DLwEjxhpAJTylDjvvMECQp
-BXNZ1p5KgE0OSMr5koI/3VXPVqAZisumgKwxBDq4/E7nGyfyeiLlBjxjzECQs2rF
-nsEWr+yWfey1p/ej3cGHo0jNbTy8c7pGz2f6OCBVRv7G8AulAg4KH3BM+co4KsBd
-QXiol24tApjgBkWHX6XOOJWCEkhzZn89AVv4vxruMCXq67nAWUvyNLlI/UGsMK3+
-gWZEyc+X717O2MgvIk+Axiporr8UIwIDAQABAoICABfD5t7wvpYtbFcOc7O2nBTX
-863YSSDKoYTNPVoEsGWBiNsDc5zwI/5gfSOnMzW56inzr6pWYFdlsO2AY6FhLiSN
-KCr0wHxBIU//RDFXKrsL5OoKjy/g7xaJC8utqpYGvr4i1v5rFlYpOosFi+5uY1rU
-0zgTsEKDqf7ThqffGodVkZzUkB2ED2Cj7PApJY0YoLLBvRvn5wSz5Cnam0r/fy/l
-EYkYsd3c6dY6bolDP/hAtVQ36xdT3V8im2YO4+Gj4U4cYVDiP4D/zvTaNCxZl0jc
-OOfnI9ZOVI2fWaiq5Zakp70AAF9ea/YI1FvWMa/XL6lQYF+7nHaFsNpxY1Rm56Nm
-a1G88axWJ2ZKrycakoStG5jD22Q1DD59s9S+FI8srAyxrwMhRbojk4SHGfCjchxt
-w/SLji7X7GoTVZIogUELi7ETBjKUxuPKfqvLm93P4mbNDAf7J9b/9GhuhypDUcCb
-ubPy0yV7iXNKsZ68J98ubigRSTc7EQFNRonNlblhkAelIbt6J8RtgBvYOmMDVOgM
-RZzVWDOgtGf8H15QUk7mTPp2F1FDx+S2YKWWF3/qs/gpTOpTJmNr0Xqqi08PDVqq
-n0eWf1rb5ouk9m103R6MwjZJcRNiODHd3n/Ml+UL4umU/yWZz3aJcfloWvfPfA5m
-s/9bU+OjR6IizZilei7dAoIBAQC/91pu5IIBlia88cZKxpKSfmns7avgdjrf5+LU
-sOVrMtJ0AOnP0xzAAshKmoRBypGlF+QJDHe4L/PPz+JFjFT3c+eEWYbnvv4Vvwq6
-z3Bl0os/Dmm1d496erQ5rTiUUk94nWjlQWkvIEypis1h9eh4ceWZHi8AwkTzLFtr
-gB+4WSASHLljIqYRre96jv3Yoyzqgy3sEf/nodQB2h2FoWLI6xQgpmEcgJt3NobJ
-7W4hnp/YRJvLiF94+xn8zy8tnW2GOeYcLP4vtteUkhmgfWaMF4vmX3FEi2Gv2k5t
-29thWe1sFVXywAn2mNmivQpxnNT+4VfugvbgaVwmV4EX0xGfAoIBAQDid16417/k
-LgBy153JupJCDUMEgLSKk19z9XkxnUTWmfBf4NHlK5VrHkBRQT/SiPHksZm4NrXr
-63xWDZhh602EF2RqZGCN+rhwU/6d2EF68pdE9kVSXnPil22KX0OXg25POzQ3cmNV
-DSgM+6lFE/FlC8s4RRmMx53wRPIYCKcit4z03eaQt8vPEQ2Q+nl7ajL1vmGmAzqC
-qaBUCPKYsdPhGkoV24FEkTYj/ABgQeoOQcEFz7BXbBqlYVKCACPPNdAxMKZESGyD
-j+fuN4JPsiQCkFvwnngBx+pz+cfrx+/xvGnbjwpHiRANKDLNikg/DYPu/FNNt0kO
-WdQ5Vh97fRb9AoIBAB9Ld9a0Me2ISPBsieMO1D5FlYcmh+qj25089GGwJPaA5oY6
-z+Mti3ZR/c0Lu5vuJig/6NEA01OXnau3PDASXJeqsYeGANQu8sUgy+MAhlZTKJgz
-BVCsGRvifRcaAitL7+C96KfXLgqcvWK2Wh0qNafNCZ0Yqpj5EryB8rVnwcQP44aQ
-YDilIWWocW4NeZyjnWLt18KwfIOKgkF7CW0Ljksq57Ea7qHy5WVd4kNxDY81iOF8
-7dkUdlwGs4nfTNWm4IANWFInk5i7uhFicAMV7m3QOOJplNGT120PrdGDEhSg0G/a
-WpSkcftdl71MDDPRWXfjzK35c73J71fg1A7KPgkCggEAGoZYtdWvSA6LRugyNvp/
-PvVGsDRS8Rgpye8qbsE42tnU6aXvb6QXMCCu79VwoDwh+ONEWesui59dRMb63h0t
-SxdiT4PKO1buW1m7LMmzga8AQLeHjwo5EIUj810NEaX3OZh5mTTwpXmRA1RGkyyu
-zWMAPmr5fx+yQ/4PTPlWeh59iACljLW6TVp1+FEeAXKC3X2a+lLOe5dai/0wTOon
-IuptWvoGvf00Firh8GRIAbt53DBjOxoSVOR9QXS6vP5UkC5eiwjgjNhCFgiyXptC
-1HoMCrbE35CGyWasH9WfKPUkgO7UCWToWGIo8SINLDVtCoTLTQL2r5Od48lOjqws
-EQKCAQEAkR/zhAu7A+pbruzOI9yxbXw83ZX4NKwNFulomM76lNZ2OQ4rnlEAsP2U
-mrHqeuD363I61AY9nb3PVW3hdhBs5Fr/RBi06Sctuj0JYChfih+EWAHmKf5/Mi1t
-94G8Z9j+G773EatgE8dCfgqidYrywIVFU1kRT3V4Jeky/WZ3UbaMnCbdScQJbhNA
-jhPEkU5ufzx+jBQxRr2A8q/NlXyRQuXOwA9d+a4zj0J2i1BwdwBkJCg3cKM+t6sv
-wCE5I1VEJWNRWMgQttMbi+nPkCUM9w5W4SweMpUyLQdva3667SDCOOnIuo8GrMlq
-kjpd0TjX4fbYHPRuLvqwl7Zjm3aUBg==
------END PRIVATE KEY-----
