--- web/ip.go
@@ -1,44 +1,0 @@
-package web
-
-import (
-	"fmt"
-	"net"
-)
-
-func GetOutboundIp() net.IP {
-	conn, err := net.Dial("udp", "8.8.8.8:80")
-	if err != nil {
-		fmt.Println(err)
-	}
-	defer conn.Close()
-
-	localAddr := conn.LocalAddr().(*net.UDPAddr)
-	return localAddr.IP
-}
-
-func GetIps() []string {
-	ips := make([]string, 0)
-	ifaces, err := net.Interfaces()
-	if err != nil {
-		panic(err)
-	}
-	for _, i := range ifaces {
-		addrs, err := i.Addrs()
-		if err != nil {
-			panic(err)
-		}
-		for _, addr := range addrs {
-			var ip net.IP
-			switch v := addr.(type) {
-			case *net.IPNet:
-				ip = v.IP
-			case *net.IPAddr:
-				ip = v.IP
-			}
-			if ip.To4() != nil && (ip.IsPrivate() || ip.IsLoopback()) {
-				ips = append(ips, ip.String())
-			}
-		}
-	}
-	return ips
-}
